Update t06_pg_algorithms.tex
The definition before did not make sense.
R is a function mapping one sample to its expected rewards. I therefore propose to remove this.

While the next line is straightforward i do not understand why we can put the two calculations together as done here. Why would the expcted value of the gradients if we sum up to (T-1) be the same as until t'?